Use bytes instead of ints and stretch out the timing
In order to save space, convert ints (16 bits) to bytes (8 bits), which
are all that are needed. Along the way, spread out the rainbow
program to provide more space between transitions.

+++ rgb_led_lamp.h
@@ -1,9 +1,11 @@
+#include "Arduino.h"
+
 typedef struct {
   // the number of millis since the arduino started when this should fire
   unsigned long fire_time;
-  int red;
-  int green;
-  int blue;
+  byte red;
+  byte green;
+  byte blue;
 } color_step;
